refactor

+++ routes/users.js
@@ -2,11 +2,7 @@ const router = require('express').Router();
 const db = require('./db');
 
 router.post('/', (req,res,next)=>{
-  db.models.User.findOne({ where: {name: req.body.name }})
-  .then( (user) => {
-    if (user) return users;
-    return db.models.User.create( { name: req.body.name })
-  })
+  db.models.User.createUser(req.body.name)
   .then( (user) => res.redirect('/'))
   .catch (next);
 });
@@ -26,9 +22,7 @@ router.delete('/:userId/user_departments/:id', (req,res,next)=>{
 })
 
 router.delete('/:id', (req,res,next)=>{
-  db.models.UserDepartment.destroy({
-    where: { userId: req.params.id }
-  })
+  db.models.UserDepartment.destroy({ where: { userId: req.params.id }})
   .then( ()=> db.models.User.destroy({ where: { id: req.params.id }}))
   .then( ()=> res.redirect('/'))
   .catch(next);
